+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- `renderPaymentForm()` function
+
 ## [1.4.14] - 2024-01-18
 
 ### Dependencies